requirements.txt
Django==5.2.3
djangorestframework==3.15.1 